+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
+## [Unreleased]
+### Added
+- Scripts that use storage for Lighttpd temporally files
+
 ## [0.9.0] - 2021-09-01
 ### Changed
 - Increase persistent logs size to 100 MB in logrotate config